修复闪存编程错误，优化编程流程

+++ User_include/Timer.h
@@ -31,12 +31,12 @@ typedef struct {
     uint8_t Tim400ms_flag;   // 400ms事件标志位
     uint8_t Tim500ms_count;  // 500ms计数
     uint8_t Tim500ms_flag;   // 500ms事件标志位
-    uint16_t Tim1s_count;    // 1s计数
-    uint16_t Tim1s_flag;     // 1s事件标志位
+    uint8_t Tim1s_count;    // 1s计数
+    uint8_t Tim1s_flag;     // 1s事件标志位
     uint16_t Tim10s_count;   // 10s计数
-    uint16_t Tim10s_flag;    // 10s事件标志位
+    uint8_t Tim10s_flag;    // 10s事件标志位
     uint16_t Tim1min_count;  // 1计数
-    uint16_t Tim1min_flag;   // 1分钟事件标志位
+    uint8_t Tim1min_flag;   // 1分钟事件标志位
 } TaskTime;
 
 #define  TaskTime_DEFAULTS  {0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0}  // 初始化参数
@@ -45,7 +45,8 @@ void SysTickConfig(void);  // 滴答定时婆渲明
 void RunSystimer(void);    // 运行计数事件标志
 void CLEAR_flag(void);     // 清除事件标志位
 
-#endif /* __TIMER_H */
+#endif
+
 
 //===========================================================================
 // No more.
